fixed bugs in methods for infoboard-application and changed logics for dto

+++ src/main/java/com/tsystems/logisticsProject/service/TruckService.java
@@ -30,6 +30,6 @@ public interface TruckService {
 
     List<Truck> findTrucksForOrder(double maxOneTimeWeight);
 
-    LinkedHashMap<String, Integer> getTrucksInfo();
+    LinkedHashMap<String, Long> getTrucksInfo();
 
 }
